Fixing UI format - Main Menu and initial changes
Moved "Configure", "Manual", and "Make a Plan" windows into a QStackedWidget within the Main Window.

This fixes the program icon on the taskbar from disappearing during regular use, as the Main Window was previously hidden.

This change requires moving all subsequent windows to open properly. This is still in progress.